Move RangeKeyComparison enum to it's own file
