+++ src/shared/src/DynamoGateway/RangeKeyComparison.ts
@@ -1,0 +1,5 @@
+enum RangeKeyComparison {
+  LessThanOrEqual
+}
+
+export default RangeKeyComparison
