+++ src/shared/src/DynamoGateway/index.ts
@@ -5,3 +5,4 @@ export { default as GetManyOptions } from "./GetManyOptions"
 export { default as UpdateOptions } from "./UpdateOptions"
 export { default as Pagination } from "./Pagination"
 export { default as IndexSearcher } from "./IndexSearcher"
+export { default as RangeKeyComparison } from "./RangeKeyComparison"
